Fixed distance computation for placing new 3d points
When placing a new point on a non-closed path, the expected behavior is
that the point would be placed a distance from the scene camera equal to
the distance from the scene camera and the last point on the spline.
However, distance was being computed by comparing the scene camera's
global coordinates and the end point's local coordinates, meaning the
computation would only be correct when the parent game object is located
at the scene origin.

Modified the distance computation to compare distances using the end
point's global position. Also added a check to pick the "end point"
based on whether the new point is getting appended or prepended.

Tested manually in scene with a PathCreator component.