Add raycast check to GetMouseWorldPosition
When building a path on top of a given scene, may intuitively expect to
click on points in the scene geometry to add path segments. However, the
path creator always creates points a pre-specified distance from the
editor camera.

Added a raycast check that looks for potential geometry the user may be
clicking on. If there is geometry close to the pre-specified distance, a
point will be generated at the ray intersection. Otherwise, we fall back
to the original behavior.

Tested manually in editor.